// app/utils/enums/announcement-placement-enum.ts
export enum AnnouncementPlacement {
  EVERY_PAGE = "every-page",
  HOME_PAGE_ONLY = "home-page-only",
  ALL_PRODUCT_PAGES = "all-product-pages",
  ALL_PRODUCTS_IN_SPECIFIC_COLLECTIONS = "all-products-in-specific-collections",
  SPECIFIC_PRODUCT_PAGE = "specific-product-page",
  ALL_COLLECTIONS_PAGE = "all-collections-page",
  SPECIFIC_COLLECTION_PAGE = "specific-collection-page",
  CUSTOM_POSITION = "custom-position",
}